+++ benches/iteration.rs
@@ -1,6 +1,9 @@
 use criterion::{black_box, Criterion, criterion_group, criterion_main};
+use memmap2::Mmap;
 
-type HiSparseBitset = hi_sparse_bitset::BitSet<hi_sparse_bitset::config::_64bit>;
+type Config = hi_sparse_bitset::config::_64bit;
+type HiSparseBitset = hi_sparse_bitset::BitSet<Config>;
+type MMapBitset<'a> = hi_sparse_bitset::DirectBitset<Config, &'a[u8], true>;
 
 fn iteration(set: &HiSparseBitset) -> u64{
     let mut s = 0;
@@ -10,13 +13,28 @@ fn iteration(set: &HiSparseBitset) -> u64{
     s
 }
 
+fn mmap_iteration(set: &MMapBitset) -> u64{
+    let mut s = 0;
+    for data in set.block_iter(){
+        s += data.len() as u64;
+    }
+    s
+}
+
 pub fn bench_iter(c: &mut Criterion) {
     let mut set: HiSparseBitset = Default::default();
-    for i in 0..3000{
-        set.insert(i*64);
+    for i in 0..30000{
+        set.insert(i*4);
     }
-    
+
+    let mut file = tempfile::tempfile().unwrap();
+    set.serialize(&mut file).unwrap();
+
+    let mmap = unsafe { Mmap::map(&file).unwrap()  };
+    let mmap_set = MMapBitset::new(&*mmap, 0).unwrap();
+
     c.bench_function("hi_sparse_bitset iter", |b| b.iter(|| iteration(black_box(&set))));
+    c.bench_function("mmap iter", |b| b.iter(|| mmap_iteration(black_box(&mmap_set))));
 }
 
 criterion_group!(benches_iter, bench_iter);