+++ src/lib.rs
@@ -159,6 +159,12 @@
 //!
 //! [serde]: https://crates.io/crates/serde
 //!
+//! # Zero-copy usage
+//!
+//! [DirectBitset] works with any byte source that can provide serialized data.
+//! You can use this with serialized data in memory-mapped file.
+//! "Almost" zero-copy.
+//!
 //! # CPU extensions
 //!
 //! Library uses `popcnt`/`count_ones` and `tzcnt`/`trailing_zeros` heavily.
@@ -191,6 +197,9 @@ mod data_block;
 
 mod bitset;
 pub use bitset::*;
+
+mod direct_bitset;
+pub use direct_bitset::*;
 
 pub mod config;
 pub mod ops;