+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.8.0
+Bitset with direct data access.
+
+### Changed
+- Serialization format changed, to support aligned reads.
+- MSRV 1.95
+
+### Added
+- `DirectBitset`.
+- `DirectDataSource` trait.
+
 ## 0.7.5
 In-place union + intersection.